Refs #20134: Separate header and sources files for TopicDataType
