+++ include/fastdds/dds/topic/TopicDataType.hpp
@@ -66,20 +66,12 @@ public:
     /**
      * @brief Constructor
      */
-    RTPS_DllAPI TopicDataType()
-        : m_typeSize(0)
-        , m_isGetKeyDefined(false)
-        , auto_fill_type_object_(true)
-        , auto_fill_type_information_(true)
-    {
-    }
+    RTPS_DllAPI TopicDataType();
 
     /**
      * @brief Destructor
      */
-    RTPS_DllAPI virtual ~TopicDataType()
-    {
-    }
+    RTPS_DllAPI virtual ~TopicDataType();
 
     /**
      * Serialize method, it should be implemented by the user, since it is abstract.
@@ -106,12 +98,7 @@ public:
     RTPS_DllAPI virtual bool serialize(
             void* data,
             fastrtps::rtps::SerializedPayload_t* payload,
-            DataRepresentationId_t data_representation)
-    {
-
-        static_cast<void>(data_representation);
-        return serialize(data, payload);
-    }
+            DataRepresentationId_t data_representation);
 
     /**
      * Deserialize method, it should be implemented by the user, since it is abstract.
@@ -142,15 +129,7 @@ public:
      */
     RTPS_DllAPI virtual std::function<uint32_t()> getSerializedSizeProvider(
             void* data,
-            DataRepresentationId_t data_representation)
-    {
-        static_cast<void>(data);
-        static_cast<void>(data_representation);
-        return []()
-               {
-                   return 0;
-               };
-    }
+            DataRepresentationId_t data_representation);
 
     /**
      * Create a Data Type.
